add `branch::remove_reference()` to do exactly that.
The most work goes into testing just to show what it really does,
and what's valid, along with some special behaviour around naming refs.
