+++ crates/but-workspace/tests/fixtures/scenario/single-branch-3-commits-no-ws-commit-more-branches.sh
@@ -1,0 +1,15 @@
+#!/usr/bin/env bash
+
+source "${BASH_SOURCE[0]%/*}/shared.sh"
+
+git init
+echo "Single commit, target, no ws commit, but ws-reference and a named segment, and branches on each commit" >.git/description
+
+commit M1
+setup_target_to_match_main
+git checkout -b A
+commit A1
+  git branch A1
+commit A2
+  git branch A2
+git checkout -b gitbutler/workspace
