+++ crates/but-workspace/tests/fixtures/scenario/single-branch-no-ws-commit-no-target.sh
@@ -8,3 +8,7 @@ echo "Single commit, no main remote/target, no ws commit, but ws-reference" >.gi
 
 commit M1
 git checkout -b gitbutler/workspace
+
+for branch in "$@"; do
+  git branch "$branch"
+done
