+++ crates/but-workspace/tests/workspace/utils.rs
@@ -13,10 +13,11 @@ pub const CONTEXT_LINES: u32 = 0;
 fn writable_scenario_inner(
     name: &str,
     creation: Creation,
+    args: impl IntoIterator<Item = impl Into<String>>,
 ) -> anyhow::Result<(gix::Repository, tempfile::TempDir)> {
     let tmp = gix_testtools::scripted_fixture_writable_with_args(
         format!("scenario/{name}.sh"),
-        None::<String>,
+        args,
         creation,
     )
     .map_err(anyhow::Error::from_boxed)?;
@@ -45,18 +46,26 @@ pub fn read_only_in_memory_scenario_named(
 }
 
 pub fn writable_scenario(name: &str) -> (gix::Repository, tempfile::TempDir) {
-    writable_scenario_inner(name, Creation::CopyFromReadOnly)
+    writable_scenario_inner(name, Creation::CopyFromReadOnly, None::<String>)
+        .expect("fixtures will yield valid repositories")
+}
+
+pub fn writable_scenario_with_args(
+    name: &str,
+    args: impl IntoIterator<Item = impl Into<String>>,
+) -> (gix::Repository, tempfile::TempDir) {
+    writable_scenario_inner(name, Creation::CopyFromReadOnly, args)
         .expect("fixtures will yield valid repositories")
 }
 
 /// It's slow because it has to re-execute the script, certain things can't be copied.
 pub fn writable_scenario_slow(name: &str) -> (gix::Repository, tempfile::TempDir) {
-    writable_scenario_inner(name, Creation::ExecuteScript)
+    writable_scenario_inner(name, Creation::ExecuteScript, None::<String>)
         .expect("fixtures will yield valid repositories")
 }
 
 pub fn writable_scenario_with_ssh_key(name: &str) -> (gix::Repository, tempfile::TempDir) {
-    let (mut repo, tmp) = writable_scenario_inner(name, Creation::CopyFromReadOnly)
+    let (mut repo, tmp) = writable_scenario_inner(name, Creation::CopyFromReadOnly, None::<String>)
         .expect("fixtures will yield valid repositories");
     let signing_key_path = repo.workdir().expect("non-bare").join("signature.key");
     assert!(
